app/file_organizer.py
#!/usr/bin/env python3
"""
Production-Grade File Organizer with Copilot Integration
"""

import os
import re
import json
import uuid
import logging
import boto3
import psycopg2
from datetime import datetime
from botocore.exceptions import ClientError
from googleapiclient.discovery import build
from google.oauth2 import service_account

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(levelname)s - %(message)s',
    handlers=[logging.StreamHandler(), logging.FileHandler('organizer.log')]
)
logger = logging.getLogger(__name__)

class FileOrganizer:
    def __init__(self, env):
        self.env = env
        self.db_conn = self.connect_db()
        self.sqs = boto3.client('sqs')
        self.s3 = boto3.client('s3')
        
        # Load Copilot-managed secrets
        self.db_secret = self.get_copilot_secret('DB_SECRET')
        self.gdrive_secret = self.get_copilot_secret('GDRIVE_CREDS')
        
        # Initialize services
        self.drive_service = build('drive', 'v3', credentials=service_account.Credentials.from_service_account_info(
            json.loads(self.gdrive_secret)
        ))

    def get_copilot_secret(self, name):
        return os.environ[name]

    def connect_db(self):
        return psycopg2.connect(
            dbname=os.environ['COPILOT_APPLICATION_NAME'],
            user=self.db_secret['username'],
            password=self.db_secret['password'],
            host=self.db_secret['host'],
            port=self.db_secret['port']
        )

    def process_files(self):
        """Main processing workflow"""
        try:
            self.migrate_db()
            files = self.fetch_gdrive_files()
            self.update_database(files)
            self.mark_duplicates()
            self.process_queues()
            self.verify_cloud_files()
            self.send_notification("Sync completed successfully")
        except Exception as e:
            self.send_notification(f"Sync failed: {str(e)}", is_error=True)
            raise

    # ... (other methods follow same pattern with environment checks)

if __name__ == "__main__":
    env = os.getenv('COPILOT_ENVIRONMENT_NAME', 'test')
    organizer = FileOrganizer(env)
    organizer.process_files()
