requirements.txt
boto3==1.26.90
google-api-python-client==2.86.0
google-auth==2.17.3
psycopg2-binary==2.9.6
requests==2.28.2
